refactor: extract runner scene render fallback
